refactor: extract dashboard process-hygiene helpers to dashboard_procs.py
